Squashed 'MvcLib/' changes from 4b327d3..c54c153
c54c153 erro ao deletar arquivo

git-subtree-dir: MvcLib
git-subtree-split: c54c15344a5afc149226c70ad307326655922643

+++ MvcLib.FsDump/DbToLocal.cs
@@ -24,10 +24,13 @@ namespace MvcLib.FsDump
             var folders = fsInfo.GetDirectories("*.*", SearchOption.AllDirectories);
             foreach (var directoryInfo in folders)
             {
-                if (directoryInfo.GetFiles("*.*", SearchOption.AllDirectories).Any())
-                    continue;
+                if (directoryInfo.Exists)
+                {
+                    if (directoryInfo.GetFiles("*.*", SearchOption.AllDirectories).Any())
+                        continue;
 
-                directoryInfo.Delete(true);
+                    directoryInfo.Delete(true);
+                }
             }
 
             if (self && !fsInfo.GetFiles("*.*", SearchOption.AllDirectories).Any())
